+++ lib/lightning/usage_tracking/client.ex
@@ -4,7 +4,7 @@ defmodule Lightning.UsageTracking.Client do
 
 
   """
-  use Tesla, only: [:post], docs: false
+  use Tesla, only: [:head, :post], docs: false
 
   alias Lightning.UsageTracking.ResponseProcessor
 
@@ -17,7 +17,17 @@ defmodule Lightning.UsageTracking.Client do
     if ResponseProcessor.successful?(response), do: :ok, else: :error
   end
 
+  def reachable?(host) do
+    build_head_client(host)
+    |> head("/")
+    |> ResponseProcessor.successful?()
+  end
+
   defp build_client(host) do
     Tesla.client([{Tesla.Middleware.BaseUrl, host}, Tesla.Middleware.JSON])
+  end
+
+  defp build_head_client(host) do
+    Tesla.client([{Tesla.Middleware.BaseUrl, host}])
   end
 end