+++ CHANGELOG.md
@@ -17,6 +17,9 @@ and this project adheres to
 
 ### Added
 
+- Allow for automatic resubmission of failed usage tracking report submissions.
+  [1789](https://github.com/OpenFn/lightning/issues/1789)
+
 ### Changed
 
 ### Fixed